api_heal_framewor/requirement.txt
pytest
requests
google-generativeai
google-genai
jsonschema